rename: SignIn 관련 이벤트 및 상태 정리
