+++ app/src/main/java/com/jukco/waitforme/ui/sign/sign_in/SignInState.kt
@@ -1,0 +1,7 @@
+package com.jukco.waitforme.ui.sign.sign_in
+
+sealed interface SignInState {
+    object Init : SignInState
+    object Loading : SignInState
+    object MovingMain : SignInState
+}
